+++ packages/astro-floors/src/components/SvgMapViewer/main.ts
