+++ packages/astro-floors/src/index.ts
@@ -1,3 +1,3 @@
 export { type Info } from 'svgmapviewer'
 
-export { SvgMapViewer } from './components/SvgMapViewer'
+export { SvgMapViewer } from './components/SvgMapViewer/SvgMapViewer'
